requirements.txt
#Python 3.8

#CLEANER #EDA
pandas>=1.5,<3.0
numpy>=1.21,<2.0

#LOADER
SQLAlchemy>=1.4,<2.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0

#API
fastapi>=0.95,<0.111
uvicorn[standard]>=0.22,<0.29
pydantic<2.0


#dev/debug
#ipython

#SQLAlchemy==2.0.31 ,psycopg2-binary==2.9.9
